% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api.R
\name{augment}
\alias{augment}
\title{Return Sample-Level Assignments}
\usage{
augment(fit)
}
\arguments{
\item{fit}{A \code{uccdf_fit} object.}
}
\value{
A data frame.
}
\description{
Return Sample-Level Assignments
}
