% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_consensus_heatmap}
\alias{plot_consensus_heatmap}
\title{Plot a Consensus Heatmap}
\usage{
plot_consensus_heatmap(fit, k = NULL, ...)
}
\arguments{
\item{fit}{A \code{uccdf_fit} object.}

\item{k}{Optional \code{K}.}

\item{...}{Additional arguments passed to \code{\link[stats:heatmap]{stats::heatmap()}}.}
}
\value{
Invisibly returns \code{fit}.
}
\description{
Plot a Consensus Heatmap
}
