% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_embedding}
\alias{plot_embedding}
\title{Plot the Latent Embedding}
\usage{
plot_embedding(fit, dims = c(1, 2), ...)
}
\arguments{
\item{fit}{A \code{uccdf_fit} object.}

\item{dims}{Two dimensions to plot.}

\item{...}{Additional arguments passed to \code{\link[graphics:plot.default]{graphics::plot()}}.}
}
\value{
Invisibly returns \code{fit}.
}
\description{
Plot the Latent Embedding
}
