% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simulate_mixed_data}
\alias{simulate_mixed_data}
\title{Simulate Mixed-Type Data}
\usage{
simulate_mixed_data(n = 120, k = 3, seed = 123)
}
\arguments{
\item{n}{Number of rows.}

\item{k}{Number of groups.}

\item{seed}{Random seed.}
}
\value{
A data frame.
}
\description{
Simulate Mixed-Type Data
}
