% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api.R
\name{select_k}
\alias{select_k}
\title{Return the K Selection Table}
\usage{
select_k(fit)
}
\arguments{
\item{fit}{A \code{uccdf_fit} object.}
}
\value{
A data frame.
}
\description{
Return the K Selection Table
}
